complete overhaul of command-line interface
command-line interface has been completely reworked to a new action
based interface that supports bash/zsh-completion via optcomplete. this
is very much a work in progress and as of this commit, the interface is
not functioning beyond printing test msgs. bash/zsh-completion is
working though :D

added optcomplete module. reason for inclusion is because I had to
modify it slightly to get it to work with bashcompinit module in zsh. I
will send a patch upstream and hopefully remove the src file and specify
optcomplete as a dependency.

refactored logging imports and renamed logger object to log in logger module

added clean.sh script for removing pyc files from all subdirectories

disabled all imports in starcluster/__init__.py for now

small cleanup and fixes here and there