createvolume uses volumecreator security group
createvolume now looks for and launches instances in the @sc-volumecreator
security group. the instances in this group are for attaching, partitioning,
and formatting EBS volumes.

added AWSError, AMIDoesNotExist, and InstanceDoesNotExist exceptions

cli now checks if help is the first argument to a given subcommand and prints
the subcommand's usage if it is

added get_image_or_none to EasyEC2, modified EasyEC2.get_image to raise
AMIDoesNotExist exception if not found, and changed
cluster._validate_instance_type_settings to use get_image_or_none

cleanup cli help and error handling