added listpublic action to list sc public amis
changed --shutdown flag to createvolume to be --no-shutdown. By default,
createvolume will shutdown its host instance to avoid the user paying
charges for an instance they may or may not no about. The option -n
allows you to keep this instance up and the vol mounted to it after the
vol has been created if needed.

added success message to createvolume action that prints the new vol id

fixed trivial bug with removevolume command. added checks to make sure
volume being deleted is in an 'available' state before attempting to
delete.

fixed bug in cluster.get_cluster when a security group has no instances.

createvolume now autodetermines device to attach to. it initially
guesses /dev/sdz and then checks if that's taken, and moves up the
alphabet in reverse until it finds a free device to attach as. add
spinner to createvolume command when waiting for host instance to come
up.