added experimental spot instance support
clustersetup.py now checks for EBS volume/partition existence when
setting up the EBS volumes on a cluster. A warning is printed if
the volume or partition is not found.

ssh.Connection.execute now checks the exit status and logs an
error if the command exit status is not 0. Passing
ignore_exit_status=True will ignore exit_status errors.

cluster.Cluster.start now sleeps for 60 seconds instead of 15 before
rechecking if the cluster is up

add DEFAULT_CLUSTER_GROUP to static.py. This will be used to provide a
default cluster template to the start/stop/sshnode/sshmaster commands