set ssh socket timeout and handle ssh exceptions
ssh.Connection class now raises ssh exceptions related to connection
and authentication errors

added wrapper exceptions for paramikos ssh exceptions.

ssh.Connection now creates the ssh socket itself instead of having paramiko
do it. This allows for tuning the timeout on the socket. timeout is now a
keyword for the Connection class  and defaults to 30 secs.