improve output of listclusters for empty clusters
listclusters will simply print N/A for each missing attribute in the output of
listclusters instead of printing nothing.

remove VOLUME_GROUP from IGNORE_GROUP setting in starcluster.static
module. I chose to put the VOLUME_GROUP back in the output of
listclusters so that users know that it exists and can be stopped
with the "stop" command.