bug fix in _load_instance_types for last commit
the fix checks that node_instance_type is a string which indicates
multiple types in the node_instance_type setting have already been
loaded. by default the node_instance_type setting comes in as a list.
the config object then parses the list and sets node_instance_type to
the "default type" (a string) and sets node_instance_types setting
to the rest of the specified types.