fix support for alternate s3 endpoints (eg EU,etc)
awsutils was passing in the same aws_region_host endpoint used for the EC2
connection which is not correct. added option aws_s3_host for specifying
alternate s3 endpoints in the config. aws_s3_host defaults to
s3.amazonaws.com if not specified

add file transfer progress to downloadimage command