raise exceptions when ssh* cmds encounter error
add CommandNotFound and InstanceNotRunning exceptions

check that instance is in a running state in ssh* cmds. this is needed
in the case of stopped EBS instances.

update InstanceDoesNotExist exception to take optional label keyword arg.
this allows methods to customize the exception message.

add check_required/has_required methods to utils.py for checking whether a
set of local system commands exists on the PATH

add SSH_TEMPLATE to static.py that contains the command to use when
kicking-off the system's ssh client.