added --mkfs-cmd option to createvolume
new --mkfs-cmd option allows specifying an alternate command to use when
formatting the new volume

createvolume command will now delete the new volume and leave the host
instance up and running in the case that an exception is thrown while
creating/partitioning/formatting the new vol. This removes the need to
handle lingering unpartitioned/formatted volumes. Leaving the host
instance up and running in the case of exception is important so that
users dont have to pay twice for an instance in order to retry creating
the new volume.