add support for unpartitioned EBS volumes
VolumeCreator class no longer partitions new EBS volumes and just
formats the entire drive instead.

DefaultClusterSetup now checks to see how many partitions are on the
volume before mounting. If there's a single partition it mounts the
partition. If there's no partitions it tries to mount the entire device.
If there's more than one partition the code notifies the user that they
must specify the partition to use in the volume's config and continues on
with mounting the next volume.