show progress bar when waiting for cluster to start
progress bar first shows the number of running nodes vs the total number
of nodes. once all nodes are 'running' the progress bar then begins to
show number of nodes with active ssh daemons vs the total number of
nodes. this gives the user a much better clue as to what's going on
behind the scenes instead of previous indeterminate spinner. hopefully
it gets the point across that just because a node is 'running' doesnt
necessarily mean SSH is alive and well...

added Cluster.wait_for_cluster method that prints a message to the
screen and waits for the nodes in the cluster to come up.

Cluster.is_cluster_up now has an optional update_progress keyword for
computing and displaying the progress.

created Cluster.refresh_interval attribute that specifies how often to
check whether nodes are up or not