remove the need for enforce_size logic
updated is_cluster_up and wait_for_cluster methods to not care about
Cluster.cluster_size *at all*. The main reason for needing the enforce_size
kwarg in the past was to handle spot instances 'appearing' out of the
blue in the instance response list when a spot request transitioned from
'open' to 'active'. The enforce_size keyword required the number of
pending/running instances to be equal to the (original) cluster_size in
order to determine that the cluster is 'up'. This meant that in the
Cluster.add_node/remove_node code the Cluster.cluster_size attribute had
to be updated so that the is_cluster_up() and wait_for_cluster() logic would
work properly.

NOW instead of relying on the cluster_size, the code simply waits for all
spot requests to become 'active', then waits for all instances to become
'running', and then for all SSH daemons to become active. In other words,
the code just "checks whats there" regardless of the original
Cluster.cluster_size used to launch the cluster. In fact, cluster_size
should now *only* be used to launch a cluster, not in the node
accounting. This has the advantage that if a spot request fails (which
is *definitely* possible) the code will simply work with what's there.