add support for config setting callbacks
config setting callbacks get executed just before loading a given
setting into the StarClusterConfig object. the callback takes the
value of the setting and returns a new or unchanged version of the
value. a setting's callback must be specified in static.py

set os.path.expanduser as the callback for the key_location setting

updated keypair test to check that the k1_location defined in the
test config is expanded with the user's home directory