auto reconnect if connections break in SSHClient
moved sftp and transport attributes to properties that will test whether
the current connections are still active or not and if not will
attempt to reconnect. if the reconnect is unsuccessful an exception is
thrown. this is useful when you restart a cluster in the development
shell.

add method to retrieve the remote server's public key