initial code for new StarCluster threadpool
this new threadpool module gives convenient access to a pool of worker
threads that can run tasks concurrently. this module handles setting all
worker threads to daemon mode and also catches and queues all exceptions
raised in worker threads so that they can be communicated back to the
main thread. implemented a better 'wait' method that allows CTRL-C to be
caught by the main thread and also raises a ThreadPoolException if any
exceptions were thrown in a worker thread. a ThreadPoolException
contains a list of all thrown exceptions.

add workerpool to install_requires in setup.py