use new threadpool to setup nodes concurrently
converted most (not all) for loops over the nodes to use the new
threadpool class. this provides a substantial performance improvement as
most setup commands are embarassingly parallel.

updated SGE install procedures to add each node manually using a
threadpool which allows more control over monitoring the status and
avoids issues with using the built-in SGE installer script on a large
number of nodes (50-100).

use node.alias instead of node.private_dns_name everywhere in
DefaultClusterSetup. This fixes an issue with dom-* names showing up in
qhost output instead of (master, node001, node002, etc). Plus, the
private_dns_names aren't really useful anyway...

updated ThreadPool.wait() to show progress bar based on number of
unfinished tasks in the pool. This automatically adds progress for a lot
of the standard setup stages without a lot of extra code

catch ThreadPoolException's in the cli module. for now if a
ThreadPoolException is thrown each thread's exeption msg, traceback, and
job name will be printed to the screen and inform the user that there is
a bug.