shutdown worker threads properly
shutdown should wait for all SuicideJobs to complete

add numtasks kword to pool.wait() that specifies how many tasks the
calling code is waiting on. this number is then used as the max value
for the progress bar used in pool.wait(). this is needed because some
jobs may have finished by the time pool.wait() checks the number of
unfinished tasks and there is no way to check how many jobs were
originally sent in the calling code because workers immediately drain
the queue when jobs are available. this numtasks kword allows the
calling code to specify how many tasks it sent and then updates the
progress based on the unfinished tasks left in the queue using this
numtasks as the maxval for the progressbar.