load tags from user-data instead
fixed the 'instance does not exist' bug yet again which happens
when one tries to operate on an instance too quickly after creating it.
when I swapped out user-data for tags in the node aliasing logic I
was relying on running add_tag calls immediately after creating the
instances/spot requests in order to assign them aliases which is a no-no
and reintroduced this bug again.

instead, assign the aliases to user-data as before but only read from
user-data when an 'alias' tag does not already exist for the node. this
ensure that the expensive user-data fetching is only done once per node.
after the first time an alias is loaded from user-data a tag is created
to prevent a refetch in the future. this required adding back in the old
_create_flat_rate_cluster/_create_spot_cluster logic to Cluster class.