add --show-ssh-status option to listclusters cmd
passing --show-ssh-status will print out the current SSH status (up or
down) for each node in a cluster. this is useful if you have a lot of
nodes and SSH is failing on a select few. using this option will pin
point the exact nodes that have issues with SSH coming up. users can
then either handle them manually or use starcluster's restart command to
reboot all nodes and try again. in the future it'd be nice if
starcluster automatically rebooted instances after not being able to
connect for some defined amount of time...

Closes #15