loadbalancer improvements
added keywords to SGELoadBalancer for enabling/disabling both dumping
and plotting the stats data.

added keywords to SGELoadBalancer for configuring the output stats file
and the plot output directory. added validation to SGELoadBalancer for
these new kwords. also added options to loadbalance command for new
kwords. If not specified, the stats file defaults to:

~/.starcluster/sge/sge-stats.csv

and the plot output directory defaults to:

~/.starcluster/sge/plots

move the responsibility of writing stats data from SGEVisualizer
to SGEStats

catch matplotlib import errors asap and raise an exception rather than
continuing on with load balancing. if the user wanted plotting, a
failure to load the visualizer should fail loudly.

minor code cleanup and refactorings