add user kwarg to SSHClient.interactive_shell()
add ability to specify user to login as when using the python-based ssh
client via SSHClient.interactive_shell(). this is the client used by the
sshmaster/sshnode commands when the local system does not have an ssh
client. up until now the --user option to sshmaster/sshnode only worked
for systems with a locally installed ssh client.