plugins: fix support for kwargs in a plugin's init
if a plugin defined any keyword arguments they were not being loaded
correctly by StarCluster due to a really dumb bug. This should be
fixed now.

moved inspect.getargspec logic to a convenience function in the utils
module

add list of any missing required settings to the PluginError
raised in Cluster.load_plugins when there are not enough settings
defined in the plugin's config

closes gh-30