stop-cmd: warn about ebs-backed spot instances
the stop command will terminate any spot cluster asking the user if they
want to "Terminate" instead of "Stop". While the message is correct,
it's a bit too subtle as pointed out by srooks10. Added a warning to
make it clearer that a spot EBS-backed cluster will be terminated, not
stopped.

closes gh-25