cluster: fix bug in create_nodes w/ spot instances
create_nodes needed to create a separate request for each alias in the
case of spot instances in order to get the node-aliasing logic to work
properly. this is ultimately because spot instances do not respect
ami_launch_index (ie all spot nodes have ami_launch_index==0). this was
ultimately causing StarCluster to assign the same alias to each node
when requesting multiple spot nodes in a single create_nodes call.