stop-cmd: *only* stop clusters - do not terminate
stop command no longer attempts to terminate a cluster if it's not
'stoppable' (non-spot, EBS-backed). terminating a cluster is now
exclusively the 'terminate' command's job. if a cluster uses a mix of
'stoppable' and 'unstoppable' (spot or s3-backed) nodes the user can
pass the new -t (--terminate-unstoppable) option to the stop command
which will stop all nodes that can be stopped and terminate the rest.
the cluster's security group will not be deleted.