static: do not automatically create config dirs
instead of automatically creating StarCluster's config dirs on import
move this logic to create_sc_config_dirs() method and update the cli and
logger modules to explicitly call it when the ~/.starcluster folder is
required. not only is this better because it's explicit, it also avoids
creating the starcluster config dirs when running setup.py given that
StarCluster's version info is retrieved from static.py via execfile in
setup.py.