node: remove user switch in add_user_to_group
this is unnecessary given that this method will be run as root by
default in the default StarCluster setup routines. if users wish to run
this method in their own plugins they're responsible for making sure
they're root before using. this will be the case as long as they don't
explicitly reconnect as a different user before using