ssh: add get_last_status() to SSHClient
convenience method for retrieving the exit status of the most recently
executed command. returns None in the case that the last command was
executed using execute_async() given that it by default backgrounds the
command using nohup and does not wait for output or the exit status.
also returns None if no command has been executed yet.