scp: call _recv_confirm after sending each file
I noticed when I was transferring a 50MB tar.gz file that not all of the
bytes were being sent - "ls -l" output on the remote host was different
than the original. After debugging _send_files using pudb, simply
setting a breakpoint and pausing the code execution allowed 'enough
time' to pass that the transfer completed successfully. Looking at scp.c
in openssh it seems they wait for some sort of response() after sending
a file so I tried this in scp.py and it seems to be working from some
limited testing.