fix bugs in keypair loading/validation
attempt to load the key_location from the config when -k is passed to
the start command and -K is not when creating a new cluster. this fixes
an issue where the default keypair specified in a cluster template's
config autoloads its key_location into the template's key_location
setting which may not pair with the key specified by the -k option.

added require_keys kwarg to cluster.get_cluster in order to toggle
whether to raise an exception or not when the cluster's keypair could
not successfully be loaded from the config. previous behavior would
simply ignore any failures in loading the keypair which is not always
the desired behavior for all use cases. if key loading fails and
require_keys=False the cluster's key_location setting is set to None to
prevent validation from incorrectly passing in the issue mentioned above.

updated listclusters, terminate, and stop commands to pass
require_keys=False when calling Cluster.get_cluster so that keys are not
strictly required to stop, list, or terminate a cluster.