cluster: call _validate_keypair if require_keys=True
This causes ClusterManager.get_cluster to raise a ClusterValidationError
if the key_location for the cluster does not exist. This allows commands
or code requiring ssh access to perform the keypair validation in a
single call to get_cluster. any key-related exceptions will bubble up to
the user as error messages.