remove pwd and grp imports to enable use on Windows
all that was used from these modules was struct_group and struct_passwd
which are simple tuple-like datastructures. put equivalent pure-Python
implementations in utils module and updated Node class to use the
'utils' versions instead. this maintains API compatibility and avoids
import errors for Windows users given that pwd and grp do not exist on
Windows

closes gh-63