ssh: handle 'Garbage packet received' exception
This exception usually occurs when an AMI has been configured to deny
SSH access for a given remote user (e.g. root) via the remote user's
authorized_keys file. This commit handles the exception and informs the
user about the underlying cause of the issue (e.g. using an unsupported
StarCluster AMI) and suggests how to fix it