loadbalance: fix _get_stats when no exec hosts
The command 'qstat -q all.q -u \* -xml' fails with exit code 1 and
doesn't return xml when there are no execution hosts in the cluster.

This breaks the case of a single node cluster where the master is *not*
an exec host. In this case jobs can still be queued on the master but
not executed until exec hosts are added - a case the loadbalancer should
definitely support. Simply removing the '-q all.q' avoids the error and
returns data for *all* jobs including jobs not yet assigned to a given
queue due to no available exec hosts. Added queues kwarg to parse_qstat
in order to filter the jobs by queue ourselves and avoid this issue with
'qstat -q <queue_name>'.

By default the loadbalancer now monitors jobs with no queue (ie
queue_name = "") in addition to jobs submitted to the default 'all.q'
queue. This allows load balancing a single node cluster with queued jobs
and no exec hosts.