loadbalance: fix _eval_add_node when no exec hosts
Handle the case where there's a single master node with no exec hosts.

In this case SGEStats.hosts is empty so the check for 'consistent' slots
across the cluster fails. This logic ultimately needs to be completely
reworked but for now update SGEStats.slots_per_host to simply return 0
in this case.

Updated _eval_add_node to avoid divide by zero when calculating
estimated time to completion when there are no exec hosts. From what I
can tell the ettc is only used to warn the user about short running
jobs so it's a good question whether this is needed at all...?