loadbalance: fix max_add calculation in _eval_add_node
The max_add calc in _eval_add_node was using len(self.stats.hosts) which
gives the number of SGE execution hosts in the cluster. Depending on
whether the master (or even other cluster nodes) are exec hosts or not
this value may or may not match the physical number cluster nodes.
Changed max_add calc to use len(self._cluster.running_nodes) instead
given that max_nodes refers to the number of physical nodes in the
cluster