start-cmd: fix bug when config is missing [global]
If the -r option is not specified the start command uses the default
value from the config. This is reasonable if [global] is defined in the
config given that the value will be set to the default if the user did
not specify. However, in the case that the user chooses *not* to define
a [global] section at all in the config cfg.globals becomes an empty
dictionary which does not contain defaults. The start command was
blindly setting refresh_interval from the [global] config if -r wasn't
specified. When no [global] section exists this results in setting
refresh_interval to None which causes a crash when waiting for the
cluster to come up.

Updated the start command to only set refresh_interval from the global
config if it's not None.

closes gh-90