ebsimage: show progress bar for CreateImage snapshot
When using ebsimage on an EBS-backed instance StarCluster will simply
use the CreateImage AWS API call rather than build the image manually.
In this case the AMI id is immediately returned and a snapshot of the
instance is assembled in the background by AWS. Updated the code in this
case to fetch the AMI immediately, determine the generated snapshot id,
and then show a progress bar on the snapshot progress rather than
use the spinner widget.

closes gh-104