volume: fix bug in sfdisk repartitioning
If necessary, the 'resizevolume' command will use sfdisk to expand the
newly resized volume's partition table assuming the original volume had
only a single partition that takes up the entire disk. Previously
StarCluster naively assumed partitions always start at 1 when
repartitioning. This is generally the case when using 'sfdisk' however if
the partition table was created interactively using 'fdisk' then the
partition will start at 2048 rather than 1 to accomodate the MBR which
isn't used on EBS/EC2. In this case the new partition did not align with
the old partition resulting in an unusable volume. Updated the code
to determine exactly where the partition starts in order to align the
new/larger partition properly with the old/smaller partition.

closes gh-100