setup.py: ignore redefinition errors w/ check.py
The setup.py script first tries to load utils from setuptools and
fallsback to distutils. In order to maintain functionality some util
functions need to be defined for distutils which causes pyflakes to
throw a redefined import error even though in this case it is warranted.
Added #pyflakes:ignore to these redefinitions to ignore the errors when
running check.py.