setup.py: work-around redef of unused import
Use the work-around fix mentioned in the following pyflakes issue:

https://github.com/kevinw/pyflakes/issues/13

This avoids the need to install a patched version of pyflakes until the
"#pyflakes:ignore" patch is included in a future release. Once the
pyflakes patch is included in a new release this commit should be
reverted to using "#pyflakes:ignore" instead.