cluster: always set key_location in master_node property
This avoids a potential issue where Cluster.nodes gets called before the
cluster's key_location has been set which caches the master node object
with a key_location of None. In this case, attempting to use the master
node's ssh object later on, even after the cluster's key_location is set
propertly, would fail due to a null key_location. To avoid this we
simply inject the cluster's key_location everytime the cached or
non-cached master_node object is fetched.