logger: remove in-memory log handler
This handler was used to store the current active session's logs in
memory so that a crash report could be easily written if necessary. For
long running processes that log a lot of output, such as the SGE load
balancer, this causes a memory leak.

Reimplemented the crash report functionality to 'grep' the session logs
from the debug.log file by PID rather than storing them in memory. This
includes fetching any lines after a matching log message that are not
formatted by the logger (e.g. exceptions)

closes gh-157