cluster: change plugins kwarg to list of plugin objs
Previously the plugin kwarg was tied to the config and the actual plugin
loading would happen at runtime by Cluster. This meant that anyone using
StarCluster as a library would have to pass a list of plugin metadata
rather than just creating a list of plugin objects. This caused an
unnecessary coupling between Cluster and StarClusterConfig.

Changed the plugins kwarg to accept a list of plugin objects. The code
will also check for the previous format and if detected will raise a
deprecation warning and load using the old convention. Moved
load_plugins code to deathrow module. In a future release it should be
merged into StarClusterConfig._load_plugins and support for the old
format via Cluster should be removed.