volume: move -F option for mkfs.ext3 to mkfs_cmd kwarg
The -F option is specific to mkfs.ext3 and should not be used
everywhere. Users will generally use something *other* than ext3 when
passing mkfs_cmd and in this case the -F option often does not apply
(e.g. mkfs.reiserfs). Moved the -F option to be apart of the default
mkfs_cmd kwarg instead.

closes gh-181