cluster: only configure tags when creating a new group
Previous code would apply tags to the cluster group regardless of
whether it was newly created or not. This was causing false-positives
when validating existing cluster groups given that any missing tags
would always be applied before validation. Fixed the code to *only*
apply tags to a security group the first time it creates the group.