config: use copy() of vol config in _load_volumes
Previously, _load_volumes would pass a reference to the original volume
settings stored in the StarClusterConfig object when parsing cluster
templates. Updated the code to use a copy() of the volume config
instead. This prevents a Cluster object or its plugin objects from
potentially modifying the original values in the StarClusterConfig
object.