setup: update dependencies to latest versions
Use >= rather than == in install_requires to avoid cases where the OS
updates packages beyond the recommended versions. In this case the code
should at least try to work with newer deps so that 1) the user can at
least run the code if it's compatible and 2) bugs can be reported
against the latest versions of the deps if it's not.

Updated README and requirements.txt with latest dependency requirements.