cluster: load volume settings from master userdata
Updated load_receipt to catch MasterDoesNotExist and check for spot
requests before reraising. If spot requests are found load_receipt will
wait for the spot requests to activate, then for instances to activate,
and then attempt to refetch the master node. This properly handles the
case where you start an *all* spot cluster and cancel the start command
early on before any instances come up and then rerun the start command
with the --no-create (-x) option. In this case the code needs to wait
for the master to 'appear' before attempting to load the original
cluster settings.

closes gh-179