start: only raise CancelledStartRequest when creating
Only catch KeyboardInterrupt around the start() call - moved the rest of
the code past start() outside of the try/except given that at this point
in the code's execution nothing is mission critical enough to warrant
catching the KeyboardInterrupt and warning the user about anything.
Updated the KeyboardInterrupt handling code around start() to only raise
CancelledStartRequest if we were in the middle of creating/configuring a
new/existing cluster.