cluster: improve userdata validation
Previously validate_userdata() was only validating the case where all
nodes are flat rate and use the same AMI and instance type. In this case
all node aliases are included in the userdata. As far as userdata *size*
is considered this is a worst case scenario. This is not a precise
calculation for all cluster configurations in general given that some
Cluster configurations (e.g. spot or multi instance type) will not
necessarily store all aliases in every node's userdata.

Updated the userdata validation to determine the maximum number of
aliases that might be stored in any given node's userdata for a given
cluster configuration (ie spot vs flat-rate vs multi-instance type)
before validating the userdata size.