tests: remove aws_credentials_validation test
This test eventually should be moved to a 'live' tests module that
requires AWS credentials, network connection, and runs using real money
and resources. Removing this test until a 'live' tests module exists
given that all default tests should be purely local and not require an
account or network access.

https://bugs.gentoo.org/show_bug.cgi?id=426528

closes gh-121