cluster: fix bug in load_receipt related to spot
Only wait for the master node to appear in load_receipt() if there are
unfulfilled spot requests (ie those without an instance_id associated
with them). This causes the code to properly reraise the
MasterDoesNotExist exception in the case that all of the cluster's spot
requests have been fulfilled and there is no master node to be found.