balancer: handle qacct error on fresh SGE install
On a fresh SGE install where no jobs have yet been submitted you get an
error about a missing accounting file. This error goes away as soon as
someone qsub's a job. Updated the code to catch this error and check
whether the accounting file is missing. If it is then the error is
ignored otherwise the error is raised.