ebsimage: only wait for snapshot if root device is in bmap
Turns out the internal metadata server is no more accurate about the
real root device than the EC2 API in the case that the AMI has an
incorrect root device name that doesnt exist in the block device.
Updated EBSImageCreator to skip showing root device snapshot progress
and simply wait on the AMI to become 'available' if the root device
listed by the EC2 API is not in the block device mapping.

closes gh-238 (for real this time ;)