node: fix root_device_name again for misconfigured AMIs
Do not use instance metadata server to fetch root device name - it
returns the same thing as the EC2 API even in the case that the root
device name is not in the block device mapping.

Updated Node.root_device_name attribute to simply handle the case that
/dev/sda1 is listed as the root device name yet only /dev/sda exists in
the block device mapping as is the case for CentOS 6.x marketplace AMIs.
All other cases simply report a warning that the root device name is
missing from the instance's block device mapping.

closes gh-238