cluster: fix instance filtering for default VPC
Amazon EC2 now enables a 'default' VPC for all new users:

http://aws.typepad.com/aws/2013/03/amazon-ec2-update-virtual-private-clouds-for-everyone.html

For existing users: any zones that haven't been used will automatically
be configured with a default VPC. Any zones that have been used before
default VPCs will continue without a default VPC.

Any security groups created in a zone with a default VPC will
automatically be VPC-enabled.

This new EC2 feature inadvertantly broke StarCluster's group-name
instance filtering given that group-name can only be used for instances
belonging to a non-VPC security group - otherwise the
instance.group-name filter must be used. StarCluster previously assumed
that it always created a non-VPC security group for each cluster. For
clusters launched in a default VPC-enabled zone this resulted in
StarCluster infinitely waiting for instances to activate due to an
incorrect instance filter.

This patch to the Cluster.nodes property simply checks whether the
cluster's security group was automatically assigned a default VPC id or
not and configures the instance filter accordingly. This patch should
work for both new and existing users regardless of whether they have a
default VPC in a given zone or not.

Thanks to @cl-s, @syoyo, and @izharw for helping to debug and fix this
issue.

closes gh-265