awsutils: add properties for platforms/default_vpc
Added EasyEC2.{supported_platforms, default_vpc} properties which return
the supported platforms (VPC or EC2+VPC) and the default vpc if any.
Updated EasyEC2.connect_to_region to clear these cached values whenever
switching regions. We're not currently using this anywhere but it's nice
to have for debugging and also could come in useful when merging VPC
support later on.