start: do not use [] for -U option's default value
Setting a default of [] instead of None results in the option appearing
in the specified_options_dict even when -U is never specified. The
specified_options_dict ultimately overwrites the config with any options in the
options dictionary that are not None. Since the default value for -U,
[], is not None this caused the userdata_scripts list from the config to
be replaced with []. This patch fixes the issue by setting the default
value to None which removes it from the specified_options_dict all
together provided the user did not specify -U.