cli: log all BaseException stacktraces to debug log
The cli module catches all exceptions that subclass StarCluster's
exception.BaseException and formats them as 'user-friendly' error
messages without stacktraces. This means if there's a bug in the code
that raises these types of exceptions it can be hard to debug where the
underlying issue comes from. Updated the code to always log the
stacktrace of these exceptions to the debug file in case we need to
debug code that raises them.