fix for DependencyViolation when deleting security group
For users with a default VPC it seems that even after the instances have
terminated if the security group is deleted too soon you get a
'DependencyViolation' error due to some unknown dependent object. Turns
out waiting a few seconds/minute usually makes this error go away.
Updated the code to continuously sleep and retry deleting the group
until it's successful. For now this will have to do until we figure out
what the dependent object is and how to wait for it to relinquish
control of the security group before deleting the group.

closes gh-270